// types/inventoryVoucher/OrderBookingTypes.d.ts
export interface OrderBookingFormData {
  orderDate: Date;
  partyId: string;
  address: string;
  mobileNo: string;
  gstin: string;
  designId: string;
  designName: string;
  designNo: string;
  wt: string;
  wtRate: string;
  polish: string;
  image: string;
  itemType: string;
  item: {
    designId: string;
    itemId: string;
    itemGl: string;
    itemName: string;
    itemShName: string;
    itemQuantity: string;
    itemRate: string;
    makingRate: string;
    itemTotal: string;
  }[]; // Add 'item' as an array of objects with 'rate' and 'total'
  totalRate: string;
  orderQuantity: string;
}
// ...
export interface OrderTableData {
  designId: number;
  designName: string;
  designNo: string;
  orderQuantity: string;
  designRate: string;
  totalRate: string;
  wt: string;
  wtRate: string;
  polish: string;
}